+++ src/game-objects/block/HiddenBlock.ts
