+++ src/game-objects/block/LargeGreenPipeShaft.ts
