+++ src/game-objects/block/StairBlock.ts
