+++ src/game-objects/item/FireFlower.ts
