+++ src/game-objects/item/Flagpole.ts
