+++ src/game-objects/item/NonSpinningCoin.ts
