+++ src/game-objects/item/RedMushroom.ts
