+++ src/game-objects/item/SpinningCoin.ts
