+++ src/game-objects/item/Star.ts
