+++ src/scenes/Game.ts
@@ -1,26 +1,26 @@
 import Phaser from 'phaser';
-import { BrickBlock } from '../game-objects/blocks/BrickBlock';
-import { FloorBlock } from '../game-objects/blocks/FloorBlock';
-import { HiddenBlock } from '../game-objects/blocks/HiddenBlock';
-import { HorizontalGreenPipe } from '../game-objects/blocks/HorizontalGreenPipe';
-import { LargeGreenPipeShaft } from '../game-objects/blocks/LargeGreenPipeShaft';
-import { LargeVerticalGreenPipe } from '../game-objects/blocks/LargeVerticalGreenPipe';
-import { MediumVerticalGreenPipe } from '../game-objects/blocks/MediumVerticalGreenPipe';
-import { QuestionBlock } from '../game-objects/blocks/QuestionBlock';
-import { SmallVerticalGreenPipe } from '../game-objects/blocks/SmallVerticalGreenPipe';
-import { StairBlock } from '../game-objects/blocks/StairBlock';
-import { UsedBlock } from '../game-objects/blocks/UsedBlock';
+import { BrickBlock } from '../game-objects/block/BrickBlock';
+import { FloorBlock } from '../game-objects/block/FloorBlock';
+import { HiddenBlock } from '../game-objects/block/HiddenBlock';
+import { HorizontalGreenPipe } from '../game-objects/block/HorizontalGreenPipe';
+import { LargeGreenPipeShaft } from '../game-objects/block/LargeGreenPipeShaft';
+import { LargeVerticalGreenPipe } from '../game-objects/block/LargeVerticalGreenPipe';
+import { MediumVerticalGreenPipe } from '../game-objects/block/MediumVerticalGreenPipe';
+import { QuestionBlock } from '../game-objects/block/QuestionBlock';
+import { SmallVerticalGreenPipe } from '../game-objects/block/SmallVerticalGreenPipe';
+import { StairBlock } from '../game-objects/block/StairBlock';
+import { UsedBlock } from '../game-objects/block/UsedBlock';
 import { Goomba } from '../game-objects/enemy/Goomba';
 import { Koopa } from '../game-objects/enemy/Koopa';
-import { FireFlower } from '../game-objects/items/FireFlower';
-import { Flagpole } from '../game-objects/items/Flagpole';
-import { GreenMushroom } from '../game-objects/items/GreenMushroom';
-import { NonSpinningCoin } from '../game-objects/items/NonSpinningCoin';
-import { RedMushroom } from '../game-objects/items/RedMushroom';
-import { SpinningCoin } from '../game-objects/items/SpinningCoin';
-import { Star } from '../game-objects/items/Star';
-import { Mario } from '../game-objects/players/Mario';
-import { IPlayer } from '../game-objects/players/types';
+import { FireFlower } from '../game-objects/item/FireFlower';
+import { Flagpole } from '../game-objects/item/Flagpole';
+import { GreenMushroom } from '../game-objects/item/GreenMushroom';
+import { NonSpinningCoin } from '../game-objects/item/NonSpinningCoin';
+import { RedMushroom } from '../game-objects/item/RedMushroom';
+import { SpinningCoin } from '../game-objects/item/SpinningCoin';
+import { Star } from '../game-objects/item/Star';
+import { Mario } from '../game-objects/player/Mario';
+import { IPlayer } from '../game-objects/player/types';
 import { Fireball } from '../game-objects/projectile/Fireball';
 import { LargeBush } from '../game-objects/scenery/LargeBush';
 import { LargeCloud } from '../game-objects/scenery/LargeCloud';
